Move things around to consolidate AFN structs
