+++ tst/EnergyPlus/unit/AirflowNetworkSolver.unit.cc
@@ -60,7 +60,6 @@
 #include "Fixtures/EnergyPlusFixture.hh"
 
 using namespace EnergyPlus;
-using namespace AirflowNetworkBalanceManager;
 using namespace AirflowNetwork;
 
 TEST_F(EnergyPlusFixture, AirflowNetwork_SolverTest_HorizontalOpening)
